+++ remat/tf2_keras/extraction_hooks.py
